requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
python-multipart==0.0.7
pandas==2.1.4
openpyxl==3.1.2
openai==1.14.0
python-dotenv==1.0.0